Create own crate for LLVM IR target LLVM 17
The main difference compared to the old dependency used is that this targets LLVM 17.

The other major differnece is that this crate is lazy, it does not try to convert all IR into a Rust friendly format at load time. Instead it thinly wraps llvm-sys and providers getters for most interesting fields.